requirements.txt
aioserial
click
PyYAML
paho-mqtt
colorlog
dynaconf